refactor: reconcile security.codecs — killbits owns it, drop CodecConfig
Both unreleased lines put a different shape on SecurityOptions.codecs against
the same native wire key: the ABI 3.2 CodecConfig (preset/enable/disable/prefer
codec-implementation selection) and the three-layer codec killbits
(CodecKillbits allow/deny). Killbits is the newer, integration-tested surface
on native 2.3.1-rc01, so it owns `codecs`. Drops CodecConfig / CodecName /
CodecPreset (codec-implementation selection + priority).

Kept the non-colliding ABI 3.2 features: SecurityOptions.ProcessTimeoutMs,
MaxEncoderThreads, and the decode/encode FormatConfig surface (DecodeFormatConfig
/ EncodeFormatConfig, moved to FormatConfig.cs). Those now bind to the canonical
killbits ImageFormat enum: gains Tiff, drops the stale Farbfeld that the killbits
native mirror (imageflow_types::killbits::ImageFormat) does not list. Format
identifiers serialize through the single ImageFormat.ToSnakeCase() mapping.
